Agent: Formatting fix Attack: Remove ToJammer Attacker: Fix GetAttacks, remove DoGetAttacks, add ToJammer Character: Remove GetCorpHangarShips, add GetAttackers, GetAttackersAsEntities, and GetJammers EVE: Fix the market stuff Jammer: Remove everything, add GetJams and ToEntity Module: Remove IsChangingAmmo and IsReloadingAmmo Station: Remove GetCorpHangarShips, fix formatting Util: Comment unnecessary tracing

+++ ISXEVEWrapperTest/Form1.cs
@@ -70,7 +70,7 @@ namespace ISXEVEWrapperTest
 
                 InnerSpace.Echo("You have " + stationList.Count + " stations with assets.");
 
-                foreach (long station in stationList)
+                foreach (int station in stationList)
                 {
                     List<Asset> assetList;
                     assetList = Ext.Me.GetAssets(station);
